requirements.txt
pyinstaller
OpenOPC-DA
polars